Removed Cash as an asset. Also connected to MySQL for transactions

+++ Entities/ChartPixel.cs
@@ -7,10 +7,10 @@ namespace RiskGame.API.Entities
 {
     public class ChartPixel
     {
-        public double Open { get; set; }
-        public double Close { get; set; }
-        public double High { get; set; }
-        public double Low { get; set; }
+        public decimal Open { get; set; }
+        public decimal Close { get; set; }
+        public decimal High { get; set; }
+        public decimal Low { get; set; }
         public int Volume { get; set; }
         public int LastFrame { get; set; }
     }
